teams-agent/pyproject.toml
[project]
name = "teams-agent"
version = "0.1.0"
description = "Microsoft Teams Agent for communicating with proprietary backend"
requires-python = ">=3.9"
dependencies = [
    "microsoft-agents-hosting-aiohttp>=0.1.0",
    "microsoft-agents-authentication-msal>=0.1.0",
    "aiohttp>=3.9.0",
    "python-dotenv>=1.0.0",
    "cryptography>=41.0.0"
]

[tool.uv]
prerelease = "allow"
dev-dependencies = []
